update and refactor on addressbook component
