+++ src/components/addressbook/AddContact/index.js
@@ -1,1 +1,1 @@
-export {default} from './add';
+export {default} from './AddContact';
