+++ src/components/addressbook/ContactList/ContactList.js
@@ -1,13 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import muiThemeable from 'material-ui/styles/muiThemeable';
-import FlatButton from 'material-ui/FlatButton';
-import { Add as AddIcon, Book as BookIcon } from 'emerald-js-ui/lib/icons3';
-import Avatar from 'material-ui/Avatar';
-import { cardSpace, tables } from 'lib/styles';
 import Immutable from 'immutable';
-import { gotoScreen } from '../../../store/wallet/screen/screenActions';
-import Address from './Contact';
+import Contact from './Contact';
 import TopBar from '../../layout/TopBar';
 
 import styles from './ContactList.scss';
@@ -17,7 +12,7 @@ const ContactList = ({ addressBook, muiTheme }) => {
   if (addressBook.size > 0) {
     list = addressBook.map((addr) => (
       <div key={addr.get('address')} style={{ border: `1px solid ${muiTheme.palette.borderColor}` }} className={styles.listItem}>
-        <Address address={addr} />
+        <Contact address={addr} />
       </div>));
   } else {
     list = (
