+++ src/components/addressbook/form.js
@@ -47,7 +47,7 @@ const Render = ({ handleSubmit, blockAddress, invalid, pristine, submitting, can
           type="text"
           validate={required} /></div>
     </Row>
-    <Row>
+    {/* <Row>
       <div style={formStyles.left}>
         <div style={formStyles.fieldName}>Description</div>
       </div>
@@ -58,7 +58,7 @@ const Render = ({ handleSubmit, blockAddress, invalid, pristine, submitting, can
           component={TextField}
           type="text"
         /></div>
-    </Row>
+    </Row> */}
     <Row>
       <div style={formStyles.left} />
       <div style={formStyles.right}>
